+++ subsys/bluetooth/mesh/access.c
@@ -8,6 +8,7 @@
 
 #include <zephyr.h>
 #include <errno.h>
+#include <stdlib.h>
 #include <sys/util.h>
 #include <sys/byteorder.h>
 
@@ -26,6 +27,26 @@
 #include "transport.h"
 #include "access.h"
 #include "foundation.h"
+#include "settings.h"
+
+/* bt_mesh_model.flags */
+enum {
+	BT_MESH_MOD_BIND_PENDING = BIT(0),
+	BT_MESH_MOD_SUB_PENDING = BIT(1),
+	BT_MESH_MOD_PUB_PENDING = BIT(2),
+	BT_MESH_MOD_NEXT_IS_PARENT = BIT(3),
+};
+
+/* Model publication information for persistent storage. */
+struct mod_pub_val {
+	uint16_t addr;
+	uint16_t key;
+	uint8_t  ttl;
+	uint8_t  retransmit;
+	uint8_t  period;
+	uint8_t  period_div:4,
+		 cred:1;
+};
 
 static const struct bt_mesh_comp *dev_comp;
 static uint16_t dev_primary_addr;
@@ -840,3 +861,392 @@ int bt_mesh_model_extend(struct bt_mesh_model *mod,
 	return 0;
 }
 #endif
+
+static int mod_set_bind(struct bt_mesh_model *mod, size_t len_rd,
+			settings_read_cb read_cb, void *cb_arg)
+{
+	ssize_t len;
+	int i;
+
+	/* Start with empty array regardless of cleared or set value */
+	for (i = 0; i < ARRAY_SIZE(mod->keys); i++) {
+		mod->keys[i] = BT_MESH_KEY_UNUSED;
+	}
+
+	if (len_rd == 0) {
+		BT_DBG("Cleared bindings for model");
+		return 0;
+	}
+
+	len = read_cb(cb_arg, mod->keys, sizeof(mod->keys));
+	if (len < 0) {
+		BT_ERR("Failed to read value (err %zd)", len);
+		return len;
+	}
+
+	BT_HEXDUMP_DBG(mod->keys, len, "val");
+
+	BT_DBG("Decoded %zu bound keys for model", len / sizeof(mod->keys[0]));
+	return 0;
+}
+
+static int mod_set_sub(struct bt_mesh_model *mod, size_t len_rd,
+		       settings_read_cb read_cb, void *cb_arg)
+{
+	ssize_t len;
+
+	/* Start with empty array regardless of cleared or set value */
+	(void)memset(mod->groups, 0, sizeof(mod->groups));
+
+	if (len_rd == 0) {
+		BT_DBG("Cleared subscriptions for model");
+		return 0;
+	}
+
+	len = read_cb(cb_arg, mod->groups, sizeof(mod->groups));
+	if (len < 0) {
+		BT_ERR("Failed to read value (err %zd)", len);
+		return len;
+	}
+
+	BT_HEXDUMP_DBG(mod->groups, len, "val");
+
+	BT_DBG("Decoded %zu subscribed group addresses for model",
+	       len / sizeof(mod->groups[0]));
+	return 0;
+}
+
+static int mod_set_pub(struct bt_mesh_model *mod, size_t len_rd,
+		       settings_read_cb read_cb, void *cb_arg)
+{
+	struct mod_pub_val pub;
+	int err;
+
+	if (!mod->pub) {
+		BT_WARN("Model has no publication context!");
+		return -EINVAL;
+	}
+
+	if (len_rd == 0) {
+		mod->pub->addr = BT_MESH_ADDR_UNASSIGNED;
+		mod->pub->key = 0U;
+		mod->pub->cred = 0U;
+		mod->pub->ttl = 0U;
+		mod->pub->period = 0U;
+		mod->pub->retransmit = 0U;
+		mod->pub->count = 0U;
+
+		BT_DBG("Cleared publication for model");
+		return 0;
+	}
+
+	if (!IS_ENABLED(CONFIG_BT_SETTINGS)) {
+		return 0;
+	}
+
+	err = bt_mesh_settings_set(read_cb, cb_arg, &pub, sizeof(pub));
+	if (err) {
+		BT_ERR("Failed to set \'model-pub\'");
+		return err;
+	}
+
+	mod->pub->addr = pub.addr;
+	mod->pub->key = pub.key;
+	mod->pub->cred = pub.cred;
+	mod->pub->ttl = pub.ttl;
+	mod->pub->period = pub.period;
+	mod->pub->retransmit = pub.retransmit;
+	mod->pub->count = 0U;
+
+	BT_DBG("Restored model publication, dst 0x%04x app_idx 0x%03x",
+	       pub.addr, pub.key);
+
+	return 0;
+}
+
+static int mod_data_set(struct bt_mesh_model *mod,
+			const char *name, size_t len_rd,
+			settings_read_cb read_cb, void *cb_arg)
+{
+	const char *next;
+
+	settings_name_next(name, &next);
+
+	if (mod->cb && mod->cb->settings_set) {
+		return mod->cb->settings_set(mod, next, len_rd,
+			read_cb, cb_arg);
+	}
+
+	return 0;
+}
+
+static int mod_set(bool vnd, const char *name, size_t len_rd,
+		   settings_read_cb read_cb, void *cb_arg)
+{
+	struct bt_mesh_model *mod;
+	uint8_t elem_idx, mod_idx;
+	uint16_t mod_key;
+	int len;
+	const char *next;
+
+	if (!name) {
+		BT_ERR("Insufficient number of arguments");
+		return -ENOENT;
+	}
+
+	mod_key = strtol(name, NULL, 16);
+	elem_idx = mod_key >> 8;
+	mod_idx = mod_key;
+
+	BT_DBG("Decoded mod_key 0x%04x as elem_idx %u mod_idx %u",
+	       mod_key, elem_idx, mod_idx);
+
+	mod = bt_mesh_model_get(vnd, elem_idx, mod_idx);
+	if (!mod) {
+		BT_ERR("Failed to get model for elem_idx %u mod_idx %u",
+		       elem_idx, mod_idx);
+		return -ENOENT;
+	}
+
+	len = settings_name_next(name, &next);
+
+	if (!next) {
+		BT_ERR("Insufficient number of arguments");
+		return -ENOENT;
+	}
+
+	if (!strncmp(next, "bind", len)) {
+		return mod_set_bind(mod, len_rd, read_cb, cb_arg);
+	}
+
+	if (!strncmp(next, "sub", len)) {
+		return mod_set_sub(mod, len_rd, read_cb, cb_arg);
+	}
+
+	if (!strncmp(next, "pub", len)) {
+		return mod_set_pub(mod, len_rd, read_cb, cb_arg);
+	}
+
+	if (!strncmp(next, "data", len)) {
+		return mod_data_set(mod, next, len_rd, read_cb, cb_arg);
+	}
+
+	BT_WARN("Unknown module key %s", next);
+	return -ENOENT;
+}
+
+static int sig_mod_set(const char *name, size_t len_rd,
+		       settings_read_cb read_cb, void *cb_arg)
+{
+	return mod_set(false, name, len_rd, read_cb, cb_arg);
+}
+
+BT_MESH_SETTINGS_DEFINE(sig_mod, "s", sig_mod_set);
+
+static int vnd_mod_set(const char *name, size_t len_rd,
+		       settings_read_cb read_cb, void *cb_arg)
+{
+	return mod_set(true, name, len_rd, read_cb, cb_arg);
+}
+
+BT_MESH_SETTINGS_DEFINE(vnd_mod, "v", vnd_mod_set);
+
+static void encode_mod_path(struct bt_mesh_model *mod, bool vnd,
+			    const char *key, char *path, size_t path_len)
+{
+	uint16_t mod_key = (((uint16_t)mod->elem_idx << 8) | mod->mod_idx);
+
+	if (vnd) {
+		snprintk(path, path_len, "bt/mesh/v/%x/%s", mod_key, key);
+	} else {
+		snprintk(path, path_len, "bt/mesh/s/%x/%s", mod_key, key);
+	}
+}
+
+static void store_pending_mod_bind(struct bt_mesh_model *mod, bool vnd)
+{
+	uint16_t keys[CONFIG_BT_MESH_MODEL_KEY_COUNT];
+	char path[20];
+	int i, count, err;
+
+	for (i = 0, count = 0; i < ARRAY_SIZE(mod->keys); i++) {
+		if (mod->keys[i] != BT_MESH_KEY_UNUSED) {
+			keys[count++] = mod->keys[i];
+			BT_DBG("model key 0x%04x", mod->keys[i]);
+		}
+	}
+
+	encode_mod_path(mod, vnd, "bind", path, sizeof(path));
+
+	if (count) {
+		err = settings_save_one(path, keys, count * sizeof(keys[0]));
+	} else {
+		err = settings_delete(path);
+	}
+
+	if (err) {
+		BT_ERR("Failed to store %s value", log_strdup(path));
+	} else {
+		BT_DBG("Stored %s value", log_strdup(path));
+	}
+}
+
+static void store_pending_mod_sub(struct bt_mesh_model *mod, bool vnd)
+{
+	uint16_t groups[CONFIG_BT_MESH_MODEL_GROUP_COUNT];
+	char path[20];
+	int i, count, err;
+
+	for (i = 0, count = 0; i < CONFIG_BT_MESH_MODEL_GROUP_COUNT; i++) {
+		if (mod->groups[i] != BT_MESH_ADDR_UNASSIGNED) {
+			groups[count++] = mod->groups[i];
+		}
+	}
+
+	encode_mod_path(mod, vnd, "sub", path, sizeof(path));
+
+	if (count) {
+		err = settings_save_one(path, groups,
+					count * sizeof(groups[0]));
+	} else {
+		err = settings_delete(path);
+	}
+
+	if (err) {
+		BT_ERR("Failed to store %s value", log_strdup(path));
+	} else {
+		BT_DBG("Stored %s value", log_strdup(path));
+	}
+}
+
+static void store_pending_mod_pub(struct bt_mesh_model *mod, bool vnd)
+{
+	struct mod_pub_val pub;
+	char path[20];
+	int err;
+
+	encode_mod_path(mod, vnd, "pub", path, sizeof(path));
+
+	if (!mod->pub || mod->pub->addr == BT_MESH_ADDR_UNASSIGNED) {
+		err = settings_delete(path);
+	} else {
+		pub.addr = mod->pub->addr;
+		pub.key = mod->pub->key;
+		pub.ttl = mod->pub->ttl;
+		pub.retransmit = mod->pub->retransmit;
+		pub.period = mod->pub->period;
+		pub.period_div = mod->pub->period_div;
+		pub.cred = mod->pub->cred;
+
+		err = settings_save_one(path, &pub, sizeof(pub));
+	}
+
+	if (err) {
+		BT_ERR("Failed to store %s value", log_strdup(path));
+	} else {
+		BT_DBG("Stored %s value", log_strdup(path));
+	}
+}
+
+static void store_pending_mod(struct bt_mesh_model *mod,
+			      struct bt_mesh_elem *elem, bool vnd,
+			      bool primary, void *user_data)
+{
+	if (!mod->flags) {
+		return;
+	}
+
+	if (mod->flags & BT_MESH_MOD_BIND_PENDING) {
+		mod->flags &= ~BT_MESH_MOD_BIND_PENDING;
+		store_pending_mod_bind(mod, vnd);
+	}
+
+	if (mod->flags & BT_MESH_MOD_SUB_PENDING) {
+		mod->flags &= ~BT_MESH_MOD_SUB_PENDING;
+		store_pending_mod_sub(mod, vnd);
+	}
+
+	if (mod->flags & BT_MESH_MOD_PUB_PENDING) {
+		mod->flags &= ~BT_MESH_MOD_PUB_PENDING;
+		store_pending_mod_pub(mod, vnd);
+	}
+}
+
+void bt_mesh_model_pending_store(void)
+{
+	bt_mesh_model_foreach(store_pending_mod, NULL);
+}
+
+void bt_mesh_model_bind_store(struct bt_mesh_model *mod)
+{
+	mod->flags |= BT_MESH_MOD_BIND_PENDING;
+	bt_mesh_settings_store_schedule(BT_MESH_SETTINGS_MOD_PENDING);
+}
+
+void bt_mesh_model_sub_store(struct bt_mesh_model *mod)
+{
+	mod->flags |= BT_MESH_MOD_SUB_PENDING;
+	bt_mesh_settings_store_schedule(BT_MESH_SETTINGS_MOD_PENDING);
+}
+
+void bt_mesh_model_pub_store(struct bt_mesh_model *mod)
+{
+	mod->flags |= BT_MESH_MOD_PUB_PENDING;
+	bt_mesh_settings_store_schedule(BT_MESH_SETTINGS_MOD_PENDING);
+}
+
+int bt_mesh_model_data_store(struct bt_mesh_model *mod, bool vnd,
+			     const char *name, const void *data,
+			     size_t data_len)
+{
+	char path[30];
+	int err;
+
+	encode_mod_path(mod, vnd, "data", path, sizeof(path));
+	if (name) {
+		strcat(path, "/");
+		strncat(path, name, 8);
+	}
+
+	if (data_len) {
+		err = settings_save_one(path, data, data_len);
+	} else {
+		err = settings_delete(path);
+	}
+
+	if (err) {
+		BT_ERR("Failed to store %s value", log_strdup(path));
+	} else {
+		BT_DBG("Stored %s value", log_strdup(path));
+	}
+	return err;
+}
+
+static void commit_mod(struct bt_mesh_model *mod, struct bt_mesh_elem *elem,
+		       bool vnd, bool primary, void *user_data)
+{
+	if (mod->pub && mod->pub->update &&
+	    mod->pub->addr != BT_MESH_ADDR_UNASSIGNED) {
+		int32_t ms = bt_mesh_model_pub_period_get(mod);
+
+		if (ms > 0) {
+			BT_DBG("Starting publish timer (period %u ms)", ms);
+			k_delayed_work_submit(&mod->pub->timer, K_MSEC(ms));
+		}
+	}
+
+	if (!IS_ENABLED(CONFIG_BT_MESH_LOW_POWER)) {
+		return;
+	}
+
+	for (int i = 0; i < ARRAY_SIZE(mod->groups); i++) {
+		if (mod->groups[i] != BT_MESH_ADDR_UNASSIGNED) {
+			bt_mesh_lpn_group_add(mod->groups[i]);
+		}
+	}
+}
+
+void bt_mesh_model_settings_commit(void)
+{
+	bt_mesh_model_foreach(commit_mod, NULL);
+}
