+++ subsys/bluetooth/mesh/access.h
@@ -6,14 +6,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* bt_mesh_model.flags */
-enum {
-	BT_MESH_MOD_BIND_PENDING = BIT(0),
-	BT_MESH_MOD_SUB_PENDING = BIT(1),
-	BT_MESH_MOD_PUB_PENDING = BIT(2),
-	BT_MESH_MOD_NEXT_IS_PARENT = BIT(3),
-};
-
 /* Tree walk return codes */
 enum bt_mesh_walk {
 	BT_MESH_WALK_STOP,
@@ -56,3 +48,9 @@ struct bt_mesh_model *bt_mesh_model_get(bool vnd, uint8_t elem_idx, uint8_t mod_
 void bt_mesh_model_recv(struct bt_mesh_net_rx *rx, struct net_buf_simple *buf);
 
 int bt_mesh_comp_register(const struct bt_mesh_comp *comp);
+
+void bt_mesh_model_pending_store(void);
+void bt_mesh_model_bind_store(struct bt_mesh_model *mod);
+void bt_mesh_model_sub_store(struct bt_mesh_model *mod);
+void bt_mesh_model_pub_store(struct bt_mesh_model *mod);
+void bt_mesh_model_settings_commit(void);
