+++ subsys/bluetooth/mesh/app_keys.h
@@ -10,27 +10,8 @@
 #include <bluetooth/mesh.h>
 #include "subnet.h"
 
-/** Mesh Application. */
-struct bt_mesh_app_key {
-	uint16_t net_idx;
-	uint16_t app_idx;
-	bool updated;
-	struct bt_mesh_app_cred {
-		uint8_t id;
-		uint8_t val[16];
-	} keys[2];
-};
-
 /** @brief Reset the app keys module. */
 void bt_mesh_app_keys_reset(void);
-
-/** @brief Get the application key with the given AppIdx.
- *
- *  @param app_idx App index.
- *
- *  @return The matching application, or NULL if the application isn't known.
- */
-struct bt_mesh_app_key *bt_mesh_app_key_get(uint16_t app_idx);
 
 /** @brief Initialize a new application key with the given parameters.
  *
@@ -77,5 +58,8 @@ uint16_t bt_mesh_app_key_find(bool dev_key, uint8_t aid,
 			      int (*cb)(struct bt_mesh_net_rx *rx,
 					const uint8_t key[16], void *cb_data),
 			      void *cb_data);
+
+/** @brief Store pending application keys in persistent storage. */
+void bt_mesh_app_key_pending_store(void);
 
 #endif /* ZEPHYR_SUBSYS_BLUETOOTH_MESH_APP_KEYS_H_ */
