+++ subsys/bluetooth/mesh/cdb.h
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2021 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+void bt_mesh_cdb_node_store(const struct bt_mesh_cdb_node *node);
+void bt_mesh_cdb_pending_store(void);
