+++ subsys/bluetooth/mesh/cfg.h
@@ -5,5 +5,6 @@
  */
 
 void bt_mesh_cfg_init(void);
+void bt_mesh_cfg_pending_store(void);
 
 bool bt_mesh_fixed_group_match(uint16_t addr);
