+++ subsys/bluetooth/mesh/main.c
@@ -108,7 +108,7 @@ int bt_mesh_provision(const uint8_t net_key[16], uint16_t net_idx,
 		memcpy(node->dev_key, dev_key, 16);
 
 		if (IS_ENABLED(CONFIG_BT_SETTINGS)) {
-			bt_mesh_store_cdb_node(node);
+			bt_mesh_cdb_node_store(node);
 		}
 	}
 
@@ -198,7 +198,7 @@ void bt_mesh_reset(void)
 	}
 
 	if (IS_ENABLED(CONFIG_BT_SETTINGS)) {
-		bt_mesh_clear_net();
+		bt_mesh_net_clear();
 	}
 
 	(void)memset(bt_mesh.dev_key, 0, sizeof(bt_mesh.dev_key));
