+++ subsys/bluetooth/mesh/rpl.h
@@ -23,8 +23,6 @@ void bt_mesh_rpl_reset(void);
 bool bt_mesh_rpl_check(struct bt_mesh_net_rx *rx,
 			struct bt_mesh_rpl **match);
 void bt_mesh_rpl_clear(void);
-struct bt_mesh_rpl *bt_mesh_rpl_find(uint16_t src);
-struct bt_mesh_rpl *bt_mesh_rpl_alloc(uint16_t src);
-void bt_mesh_rpl_foreach(bt_mesh_rpl_func_t func, void *user_data);
 void bt_mesh_rpl_update(struct bt_mesh_rpl *rpl,
 			struct bt_mesh_net_rx *rx);
+void bt_mesh_rpl_pending_store(void);
