+++ subsys/bluetooth/mesh/subnet.h
@@ -200,4 +200,13 @@ bt_mesh_subnet_has_new_key(const struct bt_mesh_subnet *sub)
 	return sub->kr_phase != BT_MESH_KR_NORMAL;
 }
 
+/** @brief Store the Subnet information in persistent storage.
+ *
+ * @param net_idx Network index to store.
+ */
+void bt_mesh_subnet_store(uint16_t net_idx);
+
+/** @brief Store the pending Subnets in persistent storage. */
+void bt_mesh_subnet_pending_store(void);
+
 #endif /* ZEPHYR_SUBSYS_BLUETOOTH_MESH_SUBNET_H_ */
