+++ subsys/bluetooth/mesh/transport.h
@@ -76,13 +76,6 @@ struct bt_mesh_ctl_friend_sub_confirm {
 	uint8_t xact;
 } __packed;
 
-struct bt_mesh_va {
-	uint16_t ref:15,
-		 changed:1;
-	uint16_t addr;
-	uint8_t  uuid[16];
-};
-
 bool bt_mesh_tx_in_progress(void);
 
 void bt_mesh_rx_reset(void);
@@ -110,12 +103,10 @@ void bt_mesh_trans_init(void);
 
 void bt_mesh_trans_reset(void);
 
-struct bt_mesh_va *bt_mesh_va_get(uint16_t index);
-
-struct bt_mesh_va *bt_mesh_va_find(uint8_t uuid[16]);
-
 uint8_t bt_mesh_va_add(uint8_t uuid[16], uint16_t *addr);
 
 uint8_t bt_mesh_va_del(uint8_t uuid[16], uint16_t *addr);
 
 uint8_t *bt_mesh_va_label_get(uint16_t addr);
+
+void bt_mesh_va_pending_store(void);
